REF: consolidate DTI/TDI/PI get_value in ExtensionIndex, associated cleanups (#31406)
