+++ pandas/core/indexes/extension.py
@@ -1,7 +1,7 @@
 """
 Shared methods for Index subclasses backed by ExtensionArray.
 """
-from typing import List
+from typing import TYPE_CHECKING, List
 
 import numpy as np
 
@@ -11,6 +11,7 @@ from pandas.util._decorators import Appender, cache_readonly
 from pandas.core.dtypes.common import (
     ensure_platform_int,
     is_dtype_equal,
+    is_integer,
     is_object_dtype,
 )
 from pandas.core.dtypes.generic import ABCSeries
@@ -19,6 +20,9 @@ from pandas.core.arrays import ExtensionArray
 from pandas.core.indexers import deprecate_ndim_indexing
 from pandas.core.indexes.base import Index
 from pandas.core.ops import get_op_result_name
+
+if TYPE_CHECKING:
+    from pandas import Series
 
 
 def inherit_from_data(name: str, delegate, cache: bool = False, wrap: bool = False):
@@ -293,3 +297,26 @@ class ExtensionIndex(Index):
         # pass copy=False because any copying will be done in the
         #  _data.astype call above
         return Index(new_values, dtype=new_values.dtype, name=self.name, copy=False)
+
+    # --------------------------------------------------------------------
+    # Indexing Methods
+
+    @Appender(Index.get_value.__doc__)
+    def get_value(self, series: "Series", key):
+        """
+        Fast lookup of value from 1-dimensional ndarray. Only use this if you
+        know what you're doing
+        """
+        try:
+            loc = self.get_loc(key)
+        except KeyError:
+            # e.g. DatetimeIndex doesn't hold integers
+            if is_integer(key) and not self.holds_integer():
+                # Fall back to positional
+                loc = key
+            else:
+                raise
+
+        return self._get_values_for_loc(series, loc)
+
+    # --------------------------------------------------------------------
